Update models.ts
The declaration of repository data has been
moved to declare it once. Now when there is
a error in try {} but not issued by JSON.parse,
the catch {} can use the repository data to generate
a new repository object. When JSON.parse crashes,
there will be a empty repository object.